Cubic Artillery
You will be given n – an integer specifying the bunkers’ maximum
capacity. Then you will be given input lines which will contain English
alphabet letters and numbers, separated by a single space. The letters
represent the bunkers and the numbers – the weapons and their capacity.
Weapons must be stored in the bunkers. The first entered bunker is the
first in which weapons are getting stored. Every weapon takes specific
capacity, equal to its number.
When a bunker overflows (it cannot contain a given weapon due to lack of
enough free capacity), it passes the weapon to the next entered bunker.