CubicsRube
You will be given n – an integer specifying the 3 dimensions of The
Rube. Only 1 integer is used for all 3 dimensions because The Rube is a
perfect cube. After that you will start receiving lines containing 4
integers separated by a single space. The first three integers will
represent a point in the 3D space, and the last integer will represent
the particles. You must bombard that cell at that point, if there is
such cell, with the given particles, adding to it – the value
corresponding to the given 4th integer. Note that the properties
specified above apply only for cells INSIDE The Rube. Also, there will
be NO cell that is hit more than 1 times by particles.

Output
• The output consists of two lines.
• On the first line print the sum of all the cells in the Rube.
• On the second line print the amount of cells which’s value has not
been changed.